pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.212"
description = "Title Hunt Backend API"
authors = ["TitleHunt <titlehunt@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
flask = "^2.2.3"
flask-migrate = "^4.0.4"
flask-wtf = "^1.1.1"
flask-mail = "^0.9.1"
flask-login = "^0.6.2"
python-dotenv = "^1.0.0"
click = "^8.1.3"
psycopg2-binary = "^2.9.5"
pydantic = "^2.4.2"
alchemical = "^0.7.1"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
python-jose = "^3.3.0"
python-multipart = "^0.0.6"
pydantic-settings = "^2.0.3"
email-validator = "^2.1.0.post1"
google-api-python-client = "1.7.8"
google-auth-httplib2 = "0.0.3"
google-auth-oauthlib = "0.4.1"
pre-commit = "^3.6.2"
wtforms-sqlalchemy = "^0.4.1"
google-auth = "^2.34.0"
boto3 = "^1.35.18"
mypy-boto3-s3 = "^1.35.16"
filetype = "^1.2.0"
moto = "^5.0.14"
pyjwt = "^2.9.0"
firebase-admin = "^6.5.0"
exponent-server-sdk = "^2.1.0"
mypy-boto3-sns = "^1.35.0"
fastapi-pagination = "^0.12.31"


[tool.poetry.group.dev.dependencies]
faker = "^18.3.1"
httpx = "^0.25.0"
types-flask-migrate = "^4.0.0.6"
types-python-jose = "^3.3.4.8"
ruff = "^0.1.8"
mypy = "^1.8.0"
pytest = "^7.4.4"
types-requests = "2.32.0.20240712"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
filterwarnings = [
  'ignore:.*werkzeug\.urls\.url.*:DeprecationWarning',
  'ignore:.*datetime\.datetime\.utcnow.*:DeprecationWarning',
]
testpaths = ["test_flask", "test_api"]

[tool.mypy]
ignore_missing_imports = true
check_untyped_defs = true
files = ["api/**/*.py", "app/**/*.py", "test_api/**/*.py", "test_flask/**/*.py"]
exclude = ["migrations/env.py", "migrations/version/*.py"]

[tool.ruff]
exclude = [
  "__pypackages__",
  "__pycache__",
  ".eggs",
  ".git",
  ".github",
  ".pytest_cache",
  ".mypy_cache",
  ".ruff_cache",
  ".venv",
  "build",
  "dist",
  "node_modules",
  "migrations",
]
line-length = 120
indent-width = 4
